adminchat 리팩토링
